requirements.txt
astroid==2.3.1
baidu-aip==2.2.17.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
frida==12.6.18
frida-tools==3.0.0
idna==2.8
isort==4.3.21
jieba==0.39
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.2
Pillow==6.2.0
prompt-toolkit==2.0.9
Pygments==2.4.2
pylint==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
wordcloud==1.5.0
wrapt==1.11.2
